Address low-risk Sonar cleanup batch across core and tests.
Simplify redundant casts/boolean expressions and modernize assertions with behavior-preserving refactors while keeping project formatting and tests green.

Made-with: Cursor

+++ jacop-core/src/main/java/org/jacop/constraints/LinearInt.java
@@ -399,8 +399,8 @@ public class LinearInt extends PrimitiveConstraint {
     // positive weights
     for (; i < pos; i++) {
       IntDomain xd = x[i].dom();
-      min = (long) xd.min() * a[i];
-      max = (long) xd.max() * a[i];
+      min = xd.min() * a[i];
+      max = xd.max() * a[i];
       f += min;
       e += max;
       variability[i] = max - min;
